requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
python-multipart==0.0.12
aiofiles==23.2.1
